fix(daemon): review follow-ups — one extractor for four platforms, honest caps, live de-bold
Four review findings on #1586's own additions, each verified before fixing.

The tool-output extractor now lives once, in `session/tool-output.ts`. Only Slack had
learned codex-acp's rawOutput shapes; telegram, discord and feishu still dropped any
non-string rawOutput, so a folded Codex command rendered there with no output — the
very symptom this PR fixed — and the three "mirrors" had silently diverged. All four
renderers import the shared copy.

The fold cap drops from 1 MiB to 256 KiB of code units. The transcript's ceiling
measures the serialized WHOLE body in UTF-8 bytes and sheds `rawOutput` entirely when
over, so a fold near the ceiling erased itself — capped output arrived as no output.
256 KiB stays under the ceiling even at 3 bytes per unit, with serialization room.

`terminal_output` is treated as the same delta stream under another key, not a
replacement: codex-acp routes every chunk through one path and renames the meta key by
client capability — the aggregate is never sent. The replacement reading would have
collapsed a command to its last chunk the day that capability is advertised.

The console's LIVE reasoning steps ('plan' lanes from the playground stream) de-bold in
`fmtStep`, so the same run reads identically while streaming and after it is re-read
from history. Stripping happens at render, never in the accumulator, so bold pairs
split across stream chunks still resolve.
